+++ src/checkpoint.c
@@ -42,6 +42,8 @@
 
 #include <string.h>
 
+#include <fti-int/defs.h> 
+#include <fti-int/types.h> 
 #include "interface.h"
 #include "ftiff.h"
 #include "api_cuda.h"
