+++ src/incremental-checkpoint.c
@@ -1,3 +1,5 @@
+#include <fti-int/defs.h> 
+#include <fti-int/types.h> 
 #include <fti-int/incremental_checkpoint.h>
 #include "interface.h"
 #include "utility.h"
