+++ src/recover.c
@@ -36,6 +36,8 @@
  *  @brief  Recovery functions for the FTI library.
  */
 
+#include <fti-int/defs.h> 
+#include <fti-int/types.h> 
 #include "interface.h"
 
 /*-------------------------------------------------------------------------*/
