+++ src/stage.c
@@ -39,6 +39,8 @@
  *  @brief  helper functions for the FTI staging feature.
  */
 
+#include <fti-int/defs.h> 
+#include <fti-int/types.h> 
 #include "interface.h"
 
 /**
